% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada-client.R
\docType{data}
\name{.kld_base_url}
\alias{.kld_base_url}
\title{Internal: base URL for Kolada v3}
\format{
An object of class \code{character} of length 1.
}
\usage{
.kld_base_url
}
\description{
Internal: base URL for Kolada v3
}
\keyword{datasets}
