% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada-client.R
\name{.kld_get}
\alias{.kld_get}
\title{Internal: GET with pagination + robust parsing}
\usage{
.kld_get(path, query = list(), page_max = Inf, pause = 0.2)
}
\arguments{
\item{path}{character, e.g. "/kpi"}

\item{query}{named list of query params}

\item{page_max}{maximum pages to pull (for tests)}

\item{pause}{seconds to pause between pages (politeness)}
}
\description{
Internal: GET with pagination + robust parsing
}
