% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada-public.R
\name{kld_municipalities}
\alias{kld_municipalities}
\title{List/search municipalities}
\usage{
kld_municipalities(query = NULL, limit = 290L)
}
\arguments{
\item{query}{optional free-text}

\item{limit}{page size hint}
}
\description{
List/search municipalities
}
