% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada-public.R
\name{kolada_available}
\alias{kolada_available}
\title{Is Kolada v3 available?
Returns TRUE/FALSE without throwing.}
\usage{
kolada_available()
}
\description{
Is Kolada v3 available?
Returns TRUE/FALSE without throwing.
}
