% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada-public.R
\name{kld_values}
\alias{kld_values}
\title{Retrieve KPI values}
\usage{
kld_values(kpi, municipality, year)
}
\arguments{
\item{kpi}{Character, KPI id (e.g. "N00951").}

\item{municipality}{Character vector of municipality ids (e.g. "0180" for Stockholm).}

\item{year}{Integer or integer vector of years to fetch.}
}
\value{
A tibble with KPI values (includes kpi id, municipality id, year, and value).
}
\description{
Queries the Kolada v2 API for values for a given KPI, municipality (or multiple), and year(s).
}
\examples{
\dontrun{
kld_values("N00951", "0180", 2020)
kld_values("N00951", c("0180","0181"), 2018:2020)
}
}
